tests/wpt/tests/html/canvas/element/text/2d.text.drawing.style.textRendering.settings.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.text.drawing.style.textRendering.settings</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.text.drawing.style.textRendering.settings</h1>
<p class="desc">Testing basic functionalities of textRendering in Canvas</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Testing basic functionalities of textRendering in Canvas");
_addTest(function(canvas, ctx) {

  // Setting textRendering with correct case.
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "optimizeSpeed";
  _assertSame(ctx.textRendering, "optimizeSpeed", "ctx.textRendering", "\"optimizeSpeed\"");

  ctx.textRendering = "optimizeLegibility";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  ctx.textRendering = "geometricPrecision";
  _assertSame(ctx.textRendering, "geometricPrecision", "ctx.textRendering", "\"geometricPrecision\"");

  ctx.textRendering = "auto";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  // Setting textRendering with incorrect case is ignored.
  ctx.textRendering = "OPtimizeSpeed";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "OPtimizELEgibility";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "GeometricPrecision";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "optimizespeed";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "optimizelegibility";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "geometricprecision";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

  ctx.textRendering = "optimizeLegibility";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  ctx.textRendering = "AUTO";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  ctx.textRendering = "Auto";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  // Setting textRendering with non-existing font variant.
  ctx.textRendering = "abcd";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  ctx.textRendering = "normal";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  ctx.textRendering = "";
  _assertSame(ctx.textRendering, "optimizeLegibility", "ctx.textRendering", "\"optimizeLegibility\"");

  ctx.textRendering = "auto";
  _assertSame(ctx.textRendering, "auto", "ctx.textRendering", "\"auto\"");

});
</script>
